+++ docs/changelog.rst
@@ -22,3 +22,6 @@ v0.1 (not yet released)
     - At least one (`at_least_one`)
     - At most one pairwise (`at_most_one_pairwise`)
     - Incremental totalizer constraints (`create_totalizer`,`increase_totalizer`,`merge_totalizer`,`extend_totalizer`) `#15 <https://github.com/lsils/bill/pull/15>`_
+* Heuristic size-reduction of UNSAT cores:
+    - Trim unsatisfiable core (`trim_core`, `trim_core_copy`) `#33 <https://github.com/lsils/bill/pull/33>`_
+    - Minimize unsatisifable core (`minimize_core`, `minimize_core_copy`) `#33 <https://github.com/lsils/bill/pull/33>`_